+++ chat_context/__init__.py
@@ -1,0 +1,3 @@
+from chat_context.chat_context import ChatContext
+
+__all__ = [ChatContext]
