+++ testbed/executor.py
@@ -26,12 +26,12 @@ def executor(code: str, run_result: list):
                        "banned_samples" : list(banned_samples[:])
                        }
             exec(compile(refined_code, "<user_input>", "exec", dont_inherit = True, optimize = 2), context)
-            # __bot_out_buff__ = context["__bot_out_buff__"]
+            # __bot_out_buff__ = chat_context["__bot_out_buff__"]
             run_result[0] = ""
             # run_result[1] 
             return
         except Exception as e:
-            # __bot_out_buff__ = context["__bot_out_buff__"]
+            # __bot_out_buff__ = chat_context["__bot_out_buff__"]
             run_result[0] = '\n'.join(list(map(str,list(traceback.extract_tb(e.__traceback__))))) + '\n' + '-|-'*10 + '\n' + str(e)
             # run_result[1] = __bot_out_buff__
             return
